requirements.txt
beautifulsoup4==4.11.1
padatious==0.4.8
pyttsx3==2.90
PyYAML==6.0
requests==2.28.1
setuptools==59.6.0
sounddevice==0.4.5
vosk==0.3.44
wikipedia==1.4.0
